refactor(web): split BookDetail.tsx into focused sub-components
Extract 8 focused sub-components from the 2773-line BookDetail.tsx:
- bookDetailUtils.ts: formatDateTime, formatDuration, formatBytes, formatTagValue
- BookDetailHeader.tsx: cover image, title, chips, lightbox (local state)
- BookDetailStatusAlerts.tsx: soft-delete/quarantine/missing-file alerts
- BookDetailActions.tsx: all action buttons Paper
- BookDetailInfoTab.tsx: metadata grid, rating widget (local state), tags
- BookDetailVersionGroup.tsx: format tray with collapsible version rows
- BookDetailFilesTab.tsx: link-version UI, format groups, iTunes, ChangeLog
- BookDetailDialogs.tsx: all dialogs, VersionsPanel, rejection history

Parent BookDetail.tsx reduced from 2773 to ~1075 lines.
Build verified: vite build succeeds with 0 new TypeScript errors.
